fix: CustomUser
Fixed CustomUser Middleware config
