+++ backend/auth_app/middleware.py
@@ -1,0 +1,11 @@
+from django.utils.timezone import now
+
+class UpdateLastActivityMiddleware:
+    def __init__(self, get_response):
+        self.get_response = get_response
+
+    def __call__(self, request):
+        response = self.get_response(request)
+        if request.user.is_authenticated:
+            request.user.updateLastActivity()
+        return response
