+++ backend/auth_app/settings.py
@@ -39,22 +39,23 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 
     'auth_app',
-    # 'auth_app.middleware.updateLastActivity',
+    'user_auth',
 ]
 
 AUTHENTICATION_BACKENDS = [
     'django.contrib.auth.backends.ModelBackend',
 ]
 
+AUTH_USER_MODEL = 'user_auth.CustomUser'
+
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
-    'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    # 'django_pudb.PudbMiddleware',
+    'auth_app.middleware.UpdateLastActivityMiddleware',
 ]
 
 LOGIN_REDIRECT_URL = 'dashboard/'
